requirements.txt
Django==3.1.4
djangorestframework==3.12.2
django-cors-headers==3.6.0
